fix:Long改为Integer

+++ base/authority/src/main/java/com/kewen/framework/base/authority/mapper/SysUserDeptMapper.java
@@ -16,5 +16,5 @@ import java.util.List;
  */
 public interface SysUserDeptMapper extends BaseMapper<SysUserDept> {
 
-    List<DeptPrimary> listUserDept(Long userId);
+    List<DeptPrimary> listUserDept(Integer userId);
 }
